Refactor tests for reusability
